Split the Handler interface out into interfaces grouped by function.
