+++ proxy-wasm-java-host/src/main/java/io/roastedroot/proxywasm/ContextHandler.java
@@ -1,0 +1,23 @@
+package io.roastedroot.proxywasm;
+
+public interface ContextHandler {
+
+    /**
+     * Set the effective context ID.
+     *
+     * @param contextID The context ID
+     * @return The result of the operation
+     */
+    default WasmResult setEffectiveContextID(int contextID) {
+        return WasmResult.UNIMPLEMENTED;
+    }
+
+    /**
+     * Indicates to the host that the plugin is done processing active context.
+     *
+     * @return The result of the operation
+     */
+    default WasmResult done() {
+        return WasmResult.NOT_FOUND;
+    }
+}
