+++ proxy-wasm-java-host/src/main/java/io/roastedroot/proxywasm/FFIHandler.java
@@ -1,0 +1,26 @@
+package io.roastedroot.proxywasm;
+
+public interface FFIHandler {
+    /**
+     * Get the function call data.
+     *
+     * @return The function call data as a byte[], or null if not available
+     */
+    default byte[] getFuncCallData() {
+        return null;
+    }
+
+    /**
+     * Set the function call data.
+     *
+     * @param data The function call data as a byte[]
+     * @return WasmResult indicating success or failure
+     */
+    default WasmResult setFuncCallData(byte[] data) {
+        return WasmResult.UNIMPLEMENTED;
+    }
+
+    default ForeignFunction getForeignFunction(String name) {
+        return null;
+    }
+}
