+++ proxy-wasm-java-host/src/main/java/io/roastedroot/proxywasm/LogHandler.java
@@ -1,0 +1,10 @@
+package io.roastedroot.proxywasm;
+
+public interface LogHandler {
+
+    default void log(LogLevel level, String message) throws WasmException {}
+
+    default LogLevel getLogLevel() throws WasmException {
+        return LogLevel.TRACE;
+    }
+}
