+++ proxy-wasm-java-host/src/main/java/io/roastedroot/proxywasm/PropertiesHandler.java
@@ -1,0 +1,14 @@
+package io.roastedroot.proxywasm;
+
+import java.util.List;
+
+public interface PropertiesHandler {
+
+    default byte[] getProperty(List<String> key) throws WasmException {
+        return null;
+    }
+
+    default WasmResult setProperty(List<String> path, byte[] value) {
+        return WasmResult.UNIMPLEMENTED;
+    }
+}
