+++ proxy-wasm-java-host/src/main/java/io/roastedroot/proxywasm/SharedData.java
@@ -1,0 +1,11 @@
+package io.roastedroot.proxywasm;
+
+public class SharedData {
+    public byte[] data;
+    public int cas;
+
+    public SharedData(byte[] data, int cas) {
+        this.data = data;
+        this.cas = cas;
+    }
+}
