+++ proxy-wasm-java-host/src/test/java/io/roastedroot/proxywasm/examples/MockSharedHandler.java
@@ -2,6 +2,7 @@ package io.roastedroot.proxywasm.examples;
 
 import io.roastedroot.proxywasm.Handler;
 import io.roastedroot.proxywasm.QueueName;
+import io.roastedroot.proxywasm.SharedData;
 import io.roastedroot.proxywasm.WasmException;
 import io.roastedroot.proxywasm.WasmResult;
 import java.util.HashMap;
